разделение кода на модули, добавление .env и .gitignore
